Pulled out the IsStillRunning behavior into a seperate service to simplify the invocationfuture
