+++ hazelcast/src/main/java/com/hazelcast/spi/impl/operationservice/impl/operations/TraceableIsStillExecutingOperation.java
@@ -5,6 +5,7 @@ import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.spi.AbstractOperation;
 import com.hazelcast.spi.UrgentSystemOperation;
 import com.hazelcast.spi.impl.NodeEngineImpl;
+import com.hazelcast.spi.impl.operationservice.impl.IsStillRunningService;
 import com.hazelcast.spi.impl.operationservice.impl.OperationServiceImpl;
 
 import java.io.IOException;
@@ -26,7 +27,8 @@ public class TraceableIsStillExecutingOperation extends AbstractOperation implem
     public void run() throws Exception {
         NodeEngineImpl nodeEngine = (NodeEngineImpl) getNodeEngine();
         OperationServiceImpl operationService = (OperationServiceImpl) nodeEngine.getOperationService();
-        boolean executing = operationService.isOperationExecuting(getCallerAddress(), getCallerUuid(),
+        IsStillRunningService isStillRunningService = operationService.getIsStillRunningService();
+        boolean executing = isStillRunningService.isOperationExecuting(getCallerAddress(), getCallerUuid(),
                 serviceName, identifier);
         getResponseHandler().sendResponse(executing);
     }
